ai_services/hoover4_research_agent/pyproject.toml
[tool.poetry]
name = "hoover4-research-agent"
version = "0.1.0"
description = "Hoover4 Research Agent Package"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "research_agent"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
langchain = ">=0.3.25,<0.4.0"
langgraph = ">=0.4.8,<0.5.0"
langchain-openai = ">=0.3.22,<0.4.0"
protobuf = "^6.31.1"
click = "^8.1.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.31.1"
httpx = "^0.28.1"
langchain-mcp-adapters = "^0.1.0"
mcp = "^1.9.2"
pydantic = "^2.5.0"
fastapi = "^0.117.1"
aiohttp = "^3.9.0"
langfuse = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^23.0"
ruff = "^0.1.0"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
